Lista de productos del Home actualizada con los cambios requeridos
