+++ cheaper/src/App.css
@@ -1,70 +1,1 @@
-body {
-  background-color: #f6fba2;
-  background-image: linear-gradient(315deg, #f6fba2 0%, #20ded3 74%);
-  font-family: Montserrat;
-}
 
- form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.supliersForm > input {
-width: 70%;
-height: 25px;
-padding: 5px;
-border-radius: 7px;
-border-style: hidden;
-}
-
-.supliersForm > input:focus {
-
-outline-color: #DDDDDD;
-box-shadow: 0 0 3pt 2pt #DDDDDD;
-}
-
-.supliersForm > label {
-  padding: 5px;
-  color: white;
-  font-size: 18px;
-  align-items: flex-start;
-}
-.supliersForm > select {
-  width: 70%;
-  height: 25px;
-  padding: 5px;
-  border-radius: 4px;
-  border-style: hidden;
-}
-.supliersForm > button{
-  padding: 10px;
-  margin: 7px;
-  color: #20ded3;
-  background-color:white;
-  border-radius: 4px;
-  border-style: hidden;
-}
-.supliersForm > button:hover{
-  background-color:#20ded3;
-  color: white;
-}
-
-h1 {
-  color: white;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-img {
-  width: auto;
-}
-.formContainer {
-  width: 100%;
-  height: 100%;
-  padding-bottom: 50px;
-}
-.logo {
-  display: flex;
-  align-items: center;
-}
