+++ cheaper/src/App.js
@@ -1,12 +1,10 @@
 import React from "react";
-import "./App.css";
 import ProductListHome from "./components/productListHome.js";
 
 function App() {
 	return (
 		<div className="App">
 			<ProductListHome />
-
 		</div>
 	);
 }
