+++ cheaper/src/components/dummyList.js
@@ -1,0 +1,68 @@
+import abarrotes from '../assets/abarrotes.jpg';//
+import aseoHogar from '../assets/aseoHogar.jpg'; //
+import cuidadoPersonal from '../assets/cuidadoPersonal.jpg';
+import carnes from '../assets/carnes.jpg';//
+import hogar from '../assets/hogar.jpg'; //
+import lacteos from '../assets/lacteos.jpg'; //
+import panaderia from '../assets/panaderia.jpg'; //
+import verduras from '../assets/verduras.jpg'; //
+import belleza from '../assets/belleza.png'; //
+
+
+const productList = [{
+  id: 'a',
+  image: panaderia,
+  name: 'Panadería fresca',
+  description: 'Son grandes',
+},
+{
+  id: 'b',
+  image: lacteos,
+  name: 'Leche',
+  description: 'Es rica',
+},
+{
+  id: 'c',
+  image: verduras,
+  name: 'Verduras',
+  description: 'Son grandes',
+},
+{
+  id: 'd',
+  image: carnes,
+  name: 'Carnes',
+  description: 'Es rica',
+},
+{
+  id: 'e',
+  image: abarrotes,
+  name: 'Abarrotes',
+  description: 'Son grandes',
+},
+{
+  id: 'f',
+  image: cuidadoPersonal,
+  name: 'Cuidado Personal',
+  description: 'Es rica',
+},
+{
+  id: 'g',
+  image: belleza,
+  name: 'Belleza',
+  description: 'Es rica',
+},
+{
+  id: 'h',
+  image: hogar,
+  name: 'Hogar',
+  description: 'Es rica',
+},
+{
+  id: 'i',
+  image: aseoHogar,
+  name: 'Abarrotes',
+  description: 'Son grandes',
+}
+]
+
+export default productList;
