+++ cheaper/src/components/productListHome.css
@@ -1,7 +1,6 @@
 body {
   background-color: #f6fba2;
   background-image: linear-gradient(315deg, #f6fba2 0%, #20ded3 74%);
-
 }
 
 @media (max-width: 768px) {
@@ -12,7 +11,6 @@ body {
 
   .gridContainer > h2 {
     font-size: 14px;
-
   }
 }
 .gridContainer {
@@ -23,7 +21,7 @@ body {
   margin: 20px;
 }
 .row {
-width: 70%;
+  width: 70%;
 }
 .card {
   display: block;
@@ -31,17 +29,15 @@ width: 70%;
   max-width: 400px;
   margin-top: 20px;
 }
-
 .card > img {
-max-width: 300px;
-max-height: 160px;
-display: block;
-margin-left: auto;
-margin-right: auto;
-border: 3px solid #ddd;
-border-radius: 7px;
+  max-width: 300px;
+  max-height: 160px;
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+  border: 3px solid #ddd;
+  border-radius: 7px;
 }
-
 .card > button{
   padding: 10px;
   margin: 7px;
@@ -54,7 +50,6 @@ border-radius: 7px;
   background-color:#20ded3;
   color: white;
 }
-
 .card > h3{
   color: white;
 }
