+++ cheaper/src/pages/pgNewProviderForm.js
@@ -1,13 +1,5 @@
 import React from "react";
 import "./App.css";
-
-// import {
-// 	BrowserRouter as Router,
-// 	Route,
-// 	Swith,
-// 	Redirect,
-// 	Link,
-// } from "react-router-dom";
 import NewProviderForm from "./components/newProviderForm.js";
 
 function App() {
